Move objective, graident, hessian to tests
