+++ graphite_maps/precision_estimation.py
@@ -76,102 +76,6 @@ def reverse_cholesky(
     return C, permutation_idx[::-1]
 
 
-def objective_function(
-    C_k: NDArray[np.floating], U: NDArray[np.floating], lambda_l2: float = 1.0
-) -> float:
-    """
-    Objective function for optimizing the affine KR map with standard Gaussian
-    reference and l2 regularized dependence.
-
-    Parameters
-    ----------
-    C_k : np.ndarray
-        The current estimate of non-zero elements in row of C-factor of
-        associate precision matrix CTC = Prec.
-    U : np.ndarray
-        The data matrix ordered according to CTC=Prec_u.
-    lambda_l2 : float, optional
-        The regularization strength for L2 regularization.
-
-    Returns
-    -------
-    float
-        The value of `objective_function`.
-    """
-    C_k = C_k.copy()
-
-    C_k[-1] = np.exp(C_k[-1])
-    Su = U.dot(C_k)
-    n, _ = U.shape
-    regularization_l2 = 0.5 * lambda_l2 * np.sum(C_k[:-1] ** 2)
-    return 0.5 * np.sum(Su**2) - n * np.log(abs(C_k[-1])) + regularization_l2
-
-
-def gradient(
-    C_k: NDArray[np.floating], U: NDArray[np.floating], lambda_l2: float = 1.0
-) -> NDArray[np.floating]:
-    """
-    Gradient of the objective function.
-
-    Parameters
-    ----------
-    C_k : np.ndarray
-        The current estimate of non-zero elements in row of C-factor of
-        associate precision matrix CTC = Prec.
-    U : np.ndarray
-        The data matrix ordered according to CTC=Prec_u.
-    lambda_l2 : float, optional
-        The regularization strength for L2 regularization.
-
-    Returns
-    -------
-    np.ndarray
-        The gradient of the objective function.
-    """
-    C_k = C_k.copy()
-
-    n, _ = U.shape
-    C_k[-1] = np.exp(C_k[-1])
-    prediction = U.dot(C_k)
-    grad = U.T.dot(prediction)
-    grad[:-1] += lambda_l2 * C_k[:-1]  # Adjust for L2 regularization
-    grad[-1] -= n / C_k[-1]  # Adjust for the -log|C_k,k| term
-    grad[-1] *= C_k[-1]  # Adjust for log-transform
-    return grad
-
-
-def hessian(
-    C_k: NDArray[np.floating], U: NDArray[np.floating], lambda_l2: float = 1.0
-) -> NDArray[np.floating]:
-    """
-    Hessian `objective_function`.
-
-    Parameters
-    ----------
-    C_k : np.ndarray
-        The current estimate of non-zero elements in row of C-factor of
-        associate precision matrix CTC = Prec.
-    U : np.ndarray
-        The data matrix ordered according to CTC=Prec_u.
-    lambda_l2 : float, optional
-        The regularization strength for L2 regularization.
-
-    Returns
-    -------
-    np.ndarray
-        The Hessian of the objective function.
-    """
-    C_k = C_k.copy()
-
-    n, _ = U.shape
-    H = U.T.dot(U)
-    np.fill_diagonal(H[:-1, :-1], H.diagonal()[:-1] + lambda_l2)  # L2-term
-    C_k[-1] = np.exp(C_k[-1])  # log-transform
-    H[-1, -1] += n / (C_k[-1] ** 2)  # Adjust for the -log|C_k,k| term
-    H[-1, -1] *= 2.0 * C_k[-1]  # log-transform adjustment
-    return H
-
-
 def solve_row_closed_form(
     U_reduced: np.ndarray,
     lambda_l2: float,
